Day 11 py
I found a bug in my intcode-implementation which day 9 didn't catch. To
solve it, I switched the memory from my past hybrid solution (initial
memory as a list and extra memory as a dict) to a 100% dict-solution.

I am not entirely sure why it didn't work before since Computer.write()
checks if the write-position is outside the length, unless it got a
negative address. It seems like dict is slower than list for memory so I
might try to re-implement lists for (initial) memory later.